 \name{admi}
 \alias{admi}
 \title{Admitted students to the seven careers of the Science Faculty}
  \usage{data(admi)}
 \format{Object of class \code{data.frame} with 445 rows and 15 columns.}
 \description{Score obtained by each of the 445 students admitted to the seven careers of the Facultad de Ciencias of the Universidad Nacional de Colombia Bogota to the first semester of 2013, and some socio demographic information:
 \describe{
    \item{carr}{a factor with the careers as its levels}
    \item{mate,cien,soci,text,imag}{score achieved in each of the areas of the admission exam}
     \item{exam}{total score of the admission exam}
   \item{gene}{gender of the admitted}
   \item{estr}{socioeconomic stratum in 3 categories}
   \item{orig}{geographic origin of the admitted}
   \item{edad}{age of the admitted in categories}
   \item{niLE}{if the admitted requires nivelation in language}
   \item{niMa}{if the admitted requires nivelation in mathematics}
   \item{estr}{socioeconomic stratum in 7 categories}
   \item{age}{age of the admitted in years}
 }
}
 \source{SIA: Academic Information System}
 \references{C.E. Pardo (2015). Estadística descriptiva multivariada. Universidad Nacional de Colombia. Facultad de Ciencias.}  
 \keyword{datasets}

