 \name{Bogota}
 \alias{Bogota}
 \title{Localities by Stratums in Bogota City}
 \description{
  Contingency Table that indicates the number of blocks of Bogota,
  in localities by stratums (DAPD 1997, p.77).
 }
 \usage{data(Bogota)}
 \format{Object whit class \code{data.frame} of 19 rows and 7 columns.}
 \source{DAPD (1997), Population, stratification and socioeconomic aspects of Bogota}
 \references{C.E. Pardo y J.E. Ortiz (2004). Analisis multivariado de datos en R. 
 Simposio de Estadistica, Cartagena Colombia.   
 }
 \keyword{datasets}

