 \name{cafe}
 \alias{cafe}
 \title{Cofee cups}
  \usage{data(cafe)}
 \format{Object of class \code{data.frame} with 12 rows and 16 columns.}
 \description{Results of the mesure of some properties of twelve coffe cups}
 \source{R. Duarte and M. Suarez and E. Moreno and P. Ortiz (1996). An\'alisis multivariado por componentes principales, de caf\'es tostados y molidos adulterados con cereales. Cenicaf\'e, 478(2):65-76}
 \references{C.E. Pardo (2023). Estadistica descriptiva multivariada. Universidad Nacional de Colombia. Facultad de Ciencias.}  
 \keyword{datasets}

