 \name{ColorAdjective}
 \docType{ColorAdjective}
 \alias{ColorAdjective}
 \title{Associations between colors and adjectives.}
 \description{
  A group of students from Nanterre University (Paris X) were presented with 
  a list of eleve colours: blue, yellow, red, white, pink, brown, purple, 
  grey, black, green and orange. Each person in the group was asked to 
  describe each color with one ore more adjectives. A final list of 89 adjectives were 
  associates with eleven colors.
 }
 \usage{data(ColorAdjective)}
 \format{Object of class data.frame with 89 rows and 11 columns.}
 \source{Jambu, M. and Lebeaux M.O. Cluster Analysis and Data Analysis. North-Holland. Amsterdam 1983.}
 \references{Fine, J. (1996), 'Iniciacion a los analisis de datos multidimensionales a partir de ejemplos',
             Notas de curso, Montevideo. }
 \keyword{datasets}
