 \name{DogBreeds}
 \alias{DogBreeds}
 \title{Dog Breeds}
 \description{
Table that describes 27 dog breeds considering their size, weight,
speed, intelligence, affectivity, aggressiveness and function.}

 \usage{data(DogBreeds)}
 \format{Object of class data.frame with 27 rows and 7 columns with the following description:

\tabular{rllll}{
      \tab VARIABLE             \tab     CATEGORIES         \tab                 \tab                \cr
 [,1] \tab Size(SIZE)            \tab     Small(sma)         \tab    Mediun(med)  \tab   Large(lar)   \cr
 [,2] \tab Weight(WEIG)          \tab     lightweight(lig)   \tab    Mediun(med)  \tab   Heavy(hea)   \cr
 [,3] \tab Speed(SPEE)           \tab     Low(low)           \tab    Mediun(med)  \tab   High(hig)    \cr
 [,4] \tab Intelligence(INTE)    \tab     Low(low)           \tab    Mediun(med)  \tab   High(hig)    \cr
 [,5] \tab Affectivity(AFFE)     \tab     Low(low)           \tab    High(hig)    \tab                \cr
 [,6] \tab aggressiveness(AGGR)  \tab     Low(low)           \tab    High(hig)    \tab                \cr
 [,7] \tab function(FUNC)        \tab     Company(com)       \tab    Hunt(hun)    \tab   Utility(uti) \cr
}

 }
 \source{Fine, J. (1996), 'Iniciacion a los analisis de datos multidimensionales a partir de ejemplos',
            Notas de clase, Montevideo.}
 \references{Brefort, A.(1982), 'Letude des races canines a partir de leurs caracteristiques qualitatives',
            HEC - Jouy en Josas}
 \keyword{datasets}

