 \name{icfes08}
 \alias{icfes08}
 \title{Departmenst by Levels of Schools in Colombia}
 \description{
  Contingency Table that classificaes the schools of Colombia by departments and level of the schools agree with the performance of its students.
 }
 \usage{data(icfes08)}
 \format{Object whit class \code{data.frame} of 29 rows and 12 columns.}
 \source{ICFES Colombia}
 \references{C.E. Pardo, M. B\'ecue and J.E. Ortiz (2013). Correspondence Analysis of Contingency Tables with Subpartitions on Rowsand Columns. Revista Colombiana de Estad\'istica, 36(1):115-144. 
 }
 \keyword{datasets}

