\encoding{latin1} 
\name{Vietnam}
 \docType{Vietnam}
 \alias{Vietnam}
 \title{Student opinions about the Vietnam War}
 \description{
The newspaper of the students of the University of
Chapel Hill (North Carolina) conducted a survey of student opinions about the
  Vietnam War in May 1967. Responses were classified by sex, year in the
  program and one of four opinions: 
\describe{
    \item{A}{defeat power of North Vietnam by widespread bombing
      and land invasion}
    \item{B}{follow the present policy}
     \item{C}{withdraw troops to strong points and open
   negotiations on elections involving the Viet Cong}
   \item{D}{immediate withdrawal of all U.S. troops}
 }
}
 \usage{data(Vietnam)}

 \format{The 3147 consulted students were classified considering the sex, year of study and chosen strategy,
         originating a  contingency table of 10 rows: M1 to M5 and F1 to F5 (the years of education are 
         from 1 to 5 and sexes are male (M) and female (F)) and 4 columns A, B, C and D. 
 }

 \source{Fine, J. (1996), 'Iniciación a los análisis de datos multidimensionales a partir de ejemplos',
             Notes of course, Montevideo
 }

 \references{Julian Faraway (2007). faraway: Functions and datasets for books by Julian Faraway,
        R package version 1.0.2, \url{http://www.maths.bath.ac.uk/} 

            }
 \keyword{datasets}
