\name{Whisky}
 \alias{Whisky}
 \title{Whisky example}
 \description{Data frame with five features of 35 whisky brands:
\describe{
    \item{price}{in Frace Francs}
    \item{malt}{proportion in percentage}
     \item{type}{by malt proportion: low, medium, pure}
   \item{aging}{in years}
   \item{taste}{mean score of a taste panel}
 }
}
 \usage{data(Whisky)}

 \source{Fine, J. (1996), 'Iniciacion a los analisis de datos multidimensionales a partir de ejemplos',
             Notes of course, Montevideo
 }            
 \keyword{datasets}
